Made a check to see if file exists before unlinking.
git-svn-id: svn://svn.freevo.org/freevo/trunk/freevo@2910 cc3e1ea1-1e01-0410-8d68-8b121e83a9d5

+++ src/plugins/icecast.py
@@ -44,7 +44,8 @@ class PluginInterface(plugin.DaemonPlugin):
             try:
                 mycmd = os.path.basename(config.ICES_CMD)
                 newm3ufile = file(os.path.join(config.FREEVO_CACHEDIR, 'changem3u.txt'), 'rb').read()
-                os.unlink(os.path.join(config.FREEVO_CACHEDIR, 'changem3u.txt'))
+                if os.path.exists(os.path.join(config.FREEVO_CACHEDIR, 'changem3u.txt')):
+                    os.unlink(os.path.join(config.FREEVO_CACHEDIR, 'changem3u.txt'))
                 os.kill(self.ices_pid, signal.SIGTERM)
                 os.waitpid(self.ices_pid, 0)
                 time.sleep(1)
