+++ src/tv/plugins/mplayer.py
@@ -9,6 +9,9 @@
 #
 # -----------------------------------------------------------------------
 # $Log$
+# Revision 1.10  2003/07/01 21:47:35  outlyer
+# Made a check to see if file exists before unlinking.
+#
 # Revision 1.9  2003/07/01 20:35:58  outlyer
 # Replaced the os.system('rm ...') calls with os.unlink()
 #
@@ -232,7 +235,7 @@ class MPlayer:
         if (os.path.isfile('./freevo_xwin') and osd.sdl_driver == 'x11' and
             config.MPLAYER_USE_WID):
             if DEBUG: print 'Got freevo_xwin and x11'
-            os.unlink('/tmp/freevo.wid')
+            if os.path.exists('/tmp/freevo.wid'): os.unlink('/tmp/freevo.wid')
             os.system('./runapp ./freevo_xwin  0 0 %s %s > /tmp/freevo.wid &' %
                       (osd.width, osd.height))
 
@@ -323,7 +326,7 @@ class MPlayer:
         while self.thread.mode == 'stop':
             time.sleep(0.05)
         print 'stopped %s app' % self.mode
-        os.unlink('/tmp/freevo.wid')
+        if os.path.exists('/tmp/freevo.wid'): os.unlink('/tmp/freevo.wid')
 
 
     def eventhandler(self, event):
@@ -407,7 +410,7 @@ class MPlayerApp(childapp.ChildApp):
         # XXX Krister testcode for proper X11 video
         if DEBUG: print 'Killing mplayer'
         util.killall('freevo_xwin')
-        os.unlink('/tmp/freevo.wid')
+        if os.path.exists ('/tmp/freevo.wid'): os.unlink('/tmp/freevo.wid')
         if config.MPLAYER_DEBUG:
             self.log_stdout.close()
             self.log_stderr.close()
