+++ src/video/plugins/mplayer.py
@@ -20,6 +20,9 @@
 #
 # -----------------------------------------------------------------------
 # $Log$
+# Revision 1.3  2003/07/01 21:47:35  outlyer
+# Made a check to see if file exists before unlinking.
+#
 # Revision 1.2  2003/07/01 20:35:58  outlyer
 # Replaced the os.system('rm ...') calls with os.unlink()
 #
@@ -335,7 +338,7 @@ class MPlayer:
         if (os.path.isfile('./freevo_xwin') and osd.sdl_driver == 'x11' and 
             config.MPLAYER_USE_WID):
             if DEBUG: print 'Got freevo_xwin and x11'
-            os.unlink('/tmp/freevo.wid')
+            if os.path.exists('/tmp/freevo.wid'): os.unlink('/tmp/freevo.wid')
             os.system('./runapp ./freevo_xwin  0 0 %s %s > /tmp/freevo.wid &' %
                       (osd.width, osd.height))
             time.sleep(1)
@@ -583,7 +586,7 @@ class MPlayerApp(childapp.ChildApp):
         # XXX Krister testcode for proper X11 video
         if DEBUG: print 'Killing mplayer'
         util.killall('freevo_xwin')
-        os.unlink('/tmp/freevo.wid')
+        if os.path.exists('/tmp/freevo.wid'): os.unlink('/tmp/freevo.wid')
 
         if config.MPLAYER_DEBUG:
             self.log_stdout.close()
